chore(blockchain-link): get rid of circular imports
