+++ packages/blockchain-link-types/src/common.ts
@@ -2,11 +2,15 @@ import type tls from 'tls';
 
 import type { OptionalKey, RequiredKey } from '@trezor/type-utils';
 
-import { BaseCurrencyCode } from './baseCurrency';
-import type { Transaction as BlockbookTransaction, VinVout } from './blockbook';
+import type {
+    AccountBalanceHistory,
+    Transaction as BlockbookTransaction,
+    FiatRatesBySymbol,
+    TokenStandard,
+    VinVout,
+} from './blockbook';
 import type {
     AddressAlias,
-    BalanceHistory,
     Token as BlockbookToken,
     TokenTransfer as BlockbookTokenTransfer,
     Utxo as BlockbookUtxo,
@@ -72,21 +76,7 @@ export interface ServerInfo {
     network: string;
 }
 
-export type TokenStandard =
-    | 'TRC10'
-    | 'ERC20'
-    | 'TRC20'
-    | 'BEP20'
-    | 'ERC721'
-    | 'TRC721'
-    | 'BEP721'
-    | 'ERC1155'
-    | 'TRC1155'
-    | 'BEP1155'
-    | 'SPL'
-    | 'SPL-2022'
-    | 'BLOCKFROST'
-    | 'STELLAR-CLASSIC';
+export type { AccountBalanceHistory, FiatRatesBySymbol, TokenStandard };
 
 export type TransferType = 'sent' | 'recv' | 'self' | 'unknown';
 
@@ -124,17 +114,6 @@ export type TransactionDetail = {
     size: number;
     totalInput: string;
     totalOutput: string;
-};
-
-export type FiatRatesBySymbol = {
-    [K in BaseCurrencyCode]?: number | undefined;
-};
-
-export type AccountBalanceHistory = Omit<
-    OptionalKey<BalanceHistory, 'sentToSelf'>,
-    'txid' | 'rates'
-> & {
-    rates: FiatRatesBySymbol;
 };
 
 export interface Transaction {
